requirements.txt
Django==3.1.
djangorestframework==3.12.1
django-extra-fields==3.0.2
psycopg2-binary==2.9.1
PyPDF2==1.26.0
django-cors-headers==3.7.0
djangorestframework-csv==2.1.1
gunicorn==20.1.0
pytest-django==4.4.0
python-docx==0.8.11
Faker==8.11.0
Pillow==8.3.2
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
python-dotenv==0.19.2
PyJWT==2.3.0
djangorestframework-simplejwt==5.0.0
facebook-sdk==3.1.0
pdfkit